refactor: move updateInventoryRecordsForNextDay to its own file
