+++ src/game-logic/reducers/index.test.js
@@ -42,29 +42,6 @@ jest.mock('../../constants', () => ({
   PRECIPITATION_CHANCE: 0,
 }))
 
-describe('updateInventoryRecordsForNextDay', () => {
-  test('records inventory records for next day', () => {
-    const {
-      todaysPurchases,
-      todaysStartingInventory,
-    } = fn.updateInventoryRecordsForNextDay({
-      inventory: [
-        { id: 'sample-item-1', quantity: 2 },
-        { id: 'sample-item-2', quantity: 5 },
-      ],
-      todaysPurchases: {
-        'sample-item-3': 3,
-      },
-    })
-
-    expect(todaysPurchases).toEqual({})
-    expect(todaysStartingInventory).toEqual({
-      'sample-item-1': 2,
-      'sample-item-2': 5,
-    })
-  })
-})
-
 describe('applyLoanInterest', () => {
   test('applies loan interest', () => {
     expect(
